+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Twigpack Changelog
 
+## 1.0.2 - 2018-09-23
+### Added
+* Added `getModuleUri()` function
+* Added `getManifestFile()` function
+
+### Changed
+* Fixed return types to allow for null
+* Code refactoring
+
 ## 1.0.1 - 2018-09-22
 ### Added
 * Better error logging if the manifest file can't be found (check `storage/logs/web.log`)